+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * [ENHANCEMENT] Add check for systemd version before attempting to query certain metrics. #1413
 * [ENHANCEMENT] Add new counters for flush requests in Linux 5.5 #1548
 * [ENHANCEMENT] The sockstat collector now exposes IPv6 statistics in addition to the existing IPv4 support. #1552
+* [ENHANCEMENT] Add unix socket support for supervisord collector #262
 * [BUGFIX] Renamed label `state` to `name` on `node_systemd_service_restart_total`. #1393
 * [BUGFIX] Fix netdev nil reference on Darwin #1414
 * [BUGFIX] Strip path.rootfs from mountpoint labels #1421